+++ examples/react-dev-demo/src/App.tsx
@@ -2,13 +2,14 @@ import './App.css';
 
 import { FlagshipProvider, FlagshipReactSdkConfig } from '@flagship.io/react-sdk';
 import React, { createContext, Dispatch, SetStateAction } from 'react';
-import { Col, Container, Row } from 'react-bootstrap';
 import { NotificationContainer, NotificationManager } from 'react-notifications';
 
 import { AppContainer } from './components/AppContainer';
 import config from './config';
 import AppHeader from './components/AppHeader';
 import QaHeader from './components/QaHeader';
+import { getLocalStorage } from './helper/utils';
+import { Container, Row, Col } from 'react-bootstrap';
 interface VisitorContext {
     [key: string]: any;
 }
@@ -41,20 +42,25 @@ export interface QA {
 export const SettingContext = createContext<AppSettings | null>(null);
 
 const App: React.FC = () => {
-    const [currentSettings, setSettings] = React.useState<SdkSettings>({
-        envId: config.envId,
-        fetchNow: config.fetchNow,
-        decisionMode: config.decisionMode as 'API',
-        pollingInterval: config.pollingInterval,
-        enableConsoleLogs: config.enableConsoleLogs,
-        enableErrorLayout: config.enableErrorLayout,
-        enableSafeMode: config.enableSafeMode,
-        timeout: config.timeout,
-        flagshipApi: config.flagshipApi,
-        apiKey: config.apiKey,
-        nodeEnv: 'production',
-        visitorData: { ...config.visitorData }
-    });
+    const initializeSettings = () => {
+        const settingsFromLocalStorage = getLocalStorage() || {};
+        return {
+            envId: config.envId,
+            fetchNow: config.fetchNow,
+            decisionMode: config.decisionMode as 'API',
+            pollingInterval: config.pollingInterval,
+            enableConsoleLogs: config.enableConsoleLogs,
+            enableErrorLayout: config.enableErrorLayout,
+            enableSafeMode: config.enableSafeMode,
+            timeout: config.timeout,
+            flagshipApi: config.flagshipApi,
+            apiKey: config.apiKey,
+            nodeEnv: 'production',
+            visitorData: { ...config.visitorData },
+            ...settingsFromLocalStorage
+        };
+    };
+    const [currentSettings, setSettings] = React.useState<SdkSettings>(initializeSettings());
     const [QA, setQA] = React.useState<QA>({
         enabled: false,
         show: {
@@ -75,6 +81,7 @@ const App: React.FC = () => {
                     timeout={currentSettings.timeout}
                     decisionMode={currentSettings.decisionMode}
                     enableSafeMode={true}
+                    initialModifications={currentSettings.initialModifications || undefined}
                     nodeEnv={currentSettings.nodeEnv}
                     visitorData={currentSettings.visitorData}
                     onInitStart={() => {
@@ -83,8 +90,12 @@ const App: React.FC = () => {
                     onInitDone={() => {
                         console.log('onInitDone - triggered');
                     }}
-                    onUpdate={({ fsModifications }) => {
-                        console.log('onUpdate - triggered');
+                    onUpdate={({ fsModifications, status: { isSdkReady } }) => {
+                        console.log(
+                            'onUpdate - triggered with modifications:' +
+                                JSON.stringify(fsModifications) +
+                                `${!isSdkReady ? ' (SDK is still not ready)' : ''}`
+                        );
                     }}
                     onBucketingSuccess={({ status }) => {
                         if (status === '200') {