+++ RELEASENOTES.md
@@ -1,6 +1,30 @@
 # Flagship REACT SDK - Release notes
 
+## ➡️ Version 2.0.8
+
+### Improvements 💪
+
+-   `onInitStart` and `onInitDone` prop is triggered in a better way.
+
+-   Flagship React SDK is forced to runs synchronously in SSR. We will update [the documentation](https://developers.flagship.io/docs/sdk/react/v2.0) to indicate features available or not in SSR.
+
+### New features 🎉
+
+-   Add universal app demo. [Click here](./examples/react-universal-demo/README.md) to see the source code.
+
+-   Add full ssr app demo. [Click here](./examples/react-ssr-demo/README.md) to see the source code.
+
+## ➡️ Version 2.0.7
+
+### Improvements 💪
+
+-   `onUpdate` prop is triggered in a better way and `status` object is giving more data, have a look to the [documentation](https://developers.flagship.io/docs/sdk/react/v2.0#useflagship-output-status) for full details.
+
+### Bug fixes 🐛
+
 ## ➡️ Version 2.0.6
+
+-   `fetchNow` prop was not true by default.
 
 ### Bug fixes 🐛
 
@@ -228,7 +252,6 @@ Be aware that `apiKey` will be mandatory in the next major release as it will us
 
 -   `useFsSynchronize` has been removed. Campaigns synchronization is now handle using `useFlagship` hook:
 
-
     -   **BEFORE**:
 
         ```jsx
@@ -244,7 +267,7 @@ Be aware that `apiKey` will be mandatory in the next major release as it will us
             <>
                 <Button
                     onClick={() => {
-                       // [...] (Update the value of "listenedValue" )
+                        // [...] (Update the value of "listenedValue" )
                     }}
                 >
                     Test
@@ -284,7 +307,7 @@ Be aware that `apiKey` will be mandatory in the next major release as it will us
                     Trigger a synchronize
                 </Button>
             </>
-        )
+        );
         ```
 
 ### Breaking changes #6 ⚠️